Adds command to record current display configuration through CCD API as a DisplayPreset. Adds array properties to DisplayPrest for CCD path info array and CCD mode info array to enable this.
Modifies DisplayConfigModeInfo to have conditional
serialization of its target mode/source mode/desktop image info child
properties. They all occupy the same memory offset in the struct but are
binary incompatible with each other, which causes the XML serialization to
fail when writing to disk.

Adds command to apply a saved DisplayPreset through the CCD API. Confirmed
to work for my display configuration scenarios -- desktop monitors only,
desktop monitors with TV, and TV only as primary.

Fixes an omission in the SetDisplayConfig PInvoke wrapper function. The
pathArray and modeInfoArray parameters were not marked as array types, but
as singular instances of the expected type.