Adds support for saving adapter device names from a DisplayConfigAdapterName struct to a DisplayPreset. This provides a mapping of Adapter ID values to a physical display adapter device name for a given DisplayPreset. This is needed because Adapter ID values are runtime-only values and may change after a reboot. This causes saved presets to stop working after a reboot. Mapping Adapter IDs to a physical display adapter device name should solve this problem by recording which adapters are being used.
Adds robust equality and hash code calculation to the LUID struct used to
hold Adapter ID values for display source information. This supports the
use of a dictionary to map LUID values to DisplayConfigAdapterName structs
as part of capturing a new preset.

Also adds console output helpers for DisplayConfigAdapterName and
DisplayConfigSourceDeviceName.