Enable tray app to handle a global hotkey.
Initial groundwork to handle register and handle a global hotkey for
eventual use in switching between two presets specified by the user. Done
by putting a message filter to pre-process the application thread's
message pump before it hits the application thread's message loop. This
was the least messy way to do it, but MSDN warns it can be a performance
hit to do so. Maybe it's because the Windows Forms framework should
generally be the one handling it?