Fixed recording/persistence of hotkey virtual key code, and display in the UI after setting a new hotkey.
Whoops, looks like modifier keycodes and the hotkey keycode are separate
fields supplied to the Windows API and are mutually incompatible in terms
of being combined into one 32-bit bitfield. So updated the data structure
to record modifer keycodes separate from the primary keycode.

Added some UI goodness to display the keycode correctly in the set hotkey
UI after the hotkey is set for a preset.