Full global hotkey support for preset switching.
Adds support for registering global hotkeys on boot and on changes to the
preset to hotkeys mapping. HotkeyMessageFilter parses the keys pressed for
a hotkey event and fires events when the hotkey event is received. Support
functions for properly displaying virtual key codes in friendly form.
Applies display presets when the mapped hotkey is pressed.